+++ EssentialFeed/EssentialFeedTests/SharedPresentation/SharedLocalizationTests.swift
@@ -12,21 +12,9 @@ final class SharedLocalizationTests: XCTestCase {
 
     func test_localizedStrings_haveKeysAndValuesForAllSupportedLocalizations() {
         let table = "Shared"
-        let presentationBundle = Bundle(for: LoadResourcePresenter<Any, DummyView>.self)
-        let localizationBundles = allLocalizationBundles(in: presentationBundle)
-        let localizedStringKeys = allLocalizedStringKeys(in: localizationBundles, table: table)
+        let bundle = Bundle(for: LoadResourcePresenter<Any, DummyView>.self)
 
-        localizationBundles.forEach { (bundle, localization) in
-            localizedStringKeys.forEach { key in
-                let localizedString = bundle.localizedString(forKey: key, value: nil, table: table)
-
-                if localizedString == key {
-                    let language = Locale.current.localizedString(forLanguageCode: localization) ?? ""
-
-                    XCTFail("Missing \(language) (\(localization)) localized string for key: '\(key)' in table: '\(table)'")
-                }
-            }
-        }
+        assertLocalizedKeyAndValuesExist(in: bundle, table)
     }
 
     private class DummyView: ResourceView {
